2.7.1 merge (#1941)
* Core naming fix (#1919)

* Fix the issue with the automated naming of cores and add a max_federate and max_broker condition to cores and brokers as well

* add additional test cases and files

* Automated formatting of repo files (#1920)


* FIx cpplint issue

* fix failing test cases due to broken function call

* Automated formatting of repo files (#1925)



* Environmental variables (#1921)

* Add support for environmental variables

* add tests for environmental variables

* Automated formatting of repo files [skip ci] (#1922)

* add a global status query (#1870)

* add a global status query

* Automated formatting of repo files (#1871)


* update date and duplicate naming race condition for cores

* Automated formatting of repo files (#1872)



* Automated formatting of repo files


* Apply suggestions from code review


* fix a few missing clears and rename a variable and method in the comms to make it clearer.

* Automated formatting of repo files (#1924)


* update docs to environment_variables


* Fix an issue with the recorder capture text file  (#1910)

* Fix an issue with the recorder capture text file storing the type in the wrong location which could cause issues on reading the file in a player

* Automated formatting of repo files (#1911)


* clang-tidy fixes

* use Json string converters instead of a custom escaping function

* Automated formatting of repo files (#1918)


* Update docs/apps/Player.md



* signal handler (#1915)

* add some test code for signal handling in the library

* Make helicsLoadSignalHandler compile

* add a signal handler callback operation and an abort operation

* Add endline for signalHandler

* add globalError call to broker  and make abort all trigger on all federates, cores and brokers, but don't destroy the objects.

* add functions and methods to the C++98 API and some docs

* fix clang-tidy issues

* Automated formatting of repo files (#1916)


* fix punctuation in docs

* change error_code argument to match consistent Styles

* Automated formatting of repo files (#1917)


* change the signal hander to have a boolean return value which controls whether the default handler is executed

* change abort all to capture all autogenerated cores and brokers.

* make changing the broker State restricted so it can't move from the error State easily.

* fix callback to be C compatible

* make the abort work for tcp cores as well

* remove the death test from the circle ci execution

* tweak the udp core timeouts and the abort calls

* Fix test cases from rebase

* fix some formatting issues

* Automated formatting of repo files (#1927)


* warning cleanup

* Apply suggestions from code review


* Update src/helics/core/CommonCore.cpp



* Automated update to SWIG generated interface files (#1930)


* Query timeouts (#1931)

* Add query timeouts and some additional checks to prevent deadlocks from queries

* add a tickForwarding method to the brokerBase to clarify intentions and simplify constructions.

* Automated formatting of repo files (#1932)


* fix cpplint warning


* add a test case and try to fix the wait for current time bug (#1933)

* add a test case and try to fix the wait for current time bug

* Automated formatting of repo files (#1935)



* Webserver tweaks (#1936)

* Fix a few minor issues with the webserver and add a default page.

* add svg index page

* update some webserver tests

* Automated formatting of repo files (#1938)


* Add css for broker server page

* Apply suggestions from code review



* Update some docs and other things related to version numbers in prepa… (#1939)

* Update some docs and other things related to version numbers in preparation for the 2.7.1 release.

* Automated formatting of repo files (#1940)


* Apply suggestions from code review

+++ tests/helics/apps/CloneTests.cpp
@@ -204,12 +204,12 @@ TEST(clone_tests, simple_clone_test_combo)
     auto pcnt = c1.pointCount();
     EXPECT_EQ(pcnt, 3U);
     if (pcnt != 3) {
-        for (int ii = 0; ii < pcnt; ++ii) {
+        for (int ii = 0; ii < static_cast<int>(pcnt); ++ii) {
             auto pv = c1.getValue(ii);
             std::cout << "Point " << ii << " source:" << std::get<1>(pv) << " at "
                       << std::get<0>(pv) << ", value: " << std::get<2>(pv) << std::endl;
         }
-        EXPECT_EQ(std::get<0>(c1.getValue(pcnt)), helics::Time{});
+        EXPECT_EQ(std::get<0>(c1.getValue(static_cast<int>(pcnt))), helics::Time{});
     }
 
     c1.saveFile("combsave.json");
